+++ src/17alasql.js
@@ -76,6 +76,9 @@ alasql.options.oracle = true;
 alasql.options.sqlite = true;
 alasql.options.orientdb = true;
 
+// for SET NOCOUNT OFF
+alasql.options.nocount = false;
+
 //alasql.options.worker = false;
 // Variables
 alasql.vars = {};
@@ -83,6 +86,8 @@ alasql.declares = {};
 
 
 alasql.prompthistory = [];
+
+alasql.plugins = {}; // If plugin already loaded
 
 alasql.from = {}; // FROM functions
 alasql.into = {}; // INTO functions