+++ src/35search.js
@@ -29,12 +29,20 @@ yy.Search.prototype.compile = function(databaseid) {
 	var self = this;
 
 	var statement = function(params,cb){
-//		console.log(31,self);
+				// console.log(31,self);
+				// console.log(32,arguments);
 		var res;
-		res = doSearch.bind(self)(dbid,params,cb);
+		doSearch.bind(self)(dbid,params,function(data){
+			// console.log(35,data);
+			res = modify(statement.query,data);
+			// console.log(37,data);
+			if(cb) res = cb(res);
+		});
+			// console.log(39,res);
 //		if(cb) res = cb(res);
 		return res;
 	};
+	statement.query = {};
 	return statement;
 };
 
@@ -519,6 +527,12 @@ function doSearch (databaseid, params, cb) {
 					// 	return processSelector(selectors,sidx+1,nest);
 					// }
 					var r = {status:1,values:nest};
+				}
+			} else if(sel.selid == 'COMMA') {
+				if(sidx+1+1 > selectors.length) {
+					return [value];
+				} else {
+					return processSelector(selectors,sidx+1,fromdata);
 				}
 			} else {
 				throw new Error('Wrong selector '+sel.selid);