+++ src/orientdb/alasql-orientdb.js
@@ -1,0 +1,5 @@
+//
+// AlaSQL OrientDB compatibility plugin
+// (c) 2015, Andrey Gershun
+//
+var yy = alasql.yy;